requirements.txt
Django==1.7.2
djangorestframework==3.0.2
wsgiref==0.1.2
